+++ src/attribute/mod.rs
@@ -3,9 +3,11 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 pub use self::attribute::{Attribute, AttributeName, AttributeNameRef};
+pub use self::attribute_type::AttributeType;
 pub use self::attribute_value::{AttributeValue, NumberList, LengthList};
 pub use self::attributes::Attributes;
 
 mod attribute;
+mod attribute_type;
 mod attribute_value;
 mod attributes;